Simplify ExploitsSingleOutput
By moving some responsibility back onto Node
